Add playoff matches. Add groundwork for generic match table rendering per event type.
