+++ pwa/package.json
@@ -23,9 +23,11 @@
     "@remix-run/serve": "2.11.1",
     "@sentry/react": "8.25.0",
     "@tanstack/react-table": "8.20.1",
+    "@types/lodash-es": "4.17.12",
     "class-variance-authority": "0.7.0",
     "clsx": "2.1.1",
     "isbot": "4.1.0",
+    "lodash-es": "4.17.21",
     "lucide-react": "0.427.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
